Use consistent node set container for all types
With the current implementation, the underlying structure of the
node sets is a vector. This is the fastest option when there is a
small number of boundary conditions, categories, etc at each level.

The NodeSet class is templated, so extending it to use alternative
storage backends is fairly simple. Eventually we will also include
the unordered_set implimentation which is likely best for large
numbers >>100 of boundary conditions at each level (seems unlikely)
and a compile time switch will be added to make this selection.