Store separate Category and BC vector for each category node
Instead of storing a set of INodes which could contain either categories
or boundary conditions, we store separate sets of both BCNodes and
CategoryNodes. This is helpful for export where the user might want
to export all of the boundary conditions together. Also, it reduces
complexity by not requiring the use of a visitor over INodes.