Encode/decode take a pointer to a Backend instance.
This feature allows us to pass in defaults, or other user information
to the encoding and decoding functions. For example, the user may want
to define that a boundary condition named pressure will be of the scalar
type, that the geometry type of a set of input files includes
dimensions and ids, that a specific input string corresponds to
some specific callable object, etc.