Remove NodeSet
The NodeSet container was removed and each level of the tree stores the categories by value. This will allow easy copying of the tree if necessary. The ModelTraits and GeometrySets are stored by shared pointer. Those are stored at the leaf, and therefore copying the pointers should not cause unexpected behaviors.