Add backend support for type mapping in YAML
The commit deals with adding backend support (mostly for YAML at the moment). The backend support allows for control over how the inports and exports take place. Most importantly, the user can define a function to map from the name in the input file to a specific type. Also, the backend support allows for sensible default configurations for the geometry types.